We have kittens moving around!
Movement of individual kittens is being done in instance
buffers, this seems like a good approach to the problem.

Refactored structs into separate files to have an easier time
messing around in the renderer.rs file.

My failed attempt at using a separate vertex buffer to move the
kittens around ended when I've started using RenderDoc, so now
we have a .cap file for this project.